mcp/requirements.txt
# RedAmon MCP Servers - Python Dependencies

# FastMCP - Fast Python MCP server framework
fastmcp>=2.0.0

# MCP SDK (core protocol implementation)
mcp>=1.0.0

# Async support
anyio>=4.0.0
httpx>=0.25.0

# SSE transport support
sse-starlette>=1.6.0
starlette>=0.27.0
uvicorn>=0.24.0

# Utility libraries
pydantic>=2.0.0
PyYAML>=6.0

# Common libraries for execute_code scripts
requests>=2.31.0
beautifulsoup4>=4.12.0
pycryptodome>=3.19.0
PyJWT>=2.8.0
paramiko>=3.4.0
impacket>=0.11.0
pwntools>=4.11.0

# HTTP parameter discovery
arjun>=2.2.7

# Browser automation (Playwright)
playwright>=1.40.0

# WebSocket terminal server
websockets>=12.0
